Added image uploading and rendering, custom actions, and keyboard offset
Added an image uploader script, a custom actions script and keyboard avoider component wrapper. Added several custom chat elements. Added selected image to the state so it can be viewed before sending. Added a button to remove images from the message. Added a notification for when the images are being uploaded.